videos/pyproject.toml
[tool.poetry]
name = "videos"
version = "0.1.0"
description = ""
authors = ["Aleksei Kurkovskii <alkurkovsk@gmail.com>"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.111.0"}
sqlalchemy = "^2.0.18"
asyncpg = "^0.27.0"
alembic = "^1.11.1"
gunicorn = "^20.1.0"
python-jose = "^3.3.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
black = "^24.4.2"
reorder-python-imports = "^3.13.0"
pydantic = "^2.7.3"
celery = "^5.4.0"
redis = "^5.0.5"
psycopg2-binary = "^2.9.9"
pylint = "^3.2.3"
faker = "^25.8.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
exclude = "migrations"
